refactor: move ProtectedClientView to components
