+++ src/app/(frontend)/client/page.tsx
@@ -1,5 +1,5 @@
 import { AuthWrapper } from '@/components/Middleware/AuthWrapper'
-import ProtectedClientView from './ProtectedClientView'
+import ProtectedClientView from '@/components/Pages/ClientDashboard/ProtectedClientView'
 
 export default async function ClientPage() {
   return (
