+++ src/components/Pages/ClientDashboard/ProtectedClientView.tsx
